psqlvector/requirements.txt
psycopg2-binary==2.9.9
numpy==1.26.4
tqdm==4.66.2
sentence_transformers